Add YouTube docutils directive
This code was originally copied from:

http://countergram.com/youtube-in-rst

and according to that website, it was released under the MIT license.

The original code created a <object> html tag and also included
additional lines to add extra parameters.

Those extra parameters were removed and the <object> tag is replaced
by the new <iframe> tag.

To use the youtube directive:
Get the embed code from YouTube - pretend it's 1234567890
In the restructured text source file add a line similar to:

.. youtube:: 1234567890
   width=600
   height=400