explicit conversion lang -> locale (neutral in linux)
In functions that need to pass a locale, we prepare the (locale_name, encoding) in pylocale

When a function receives a 'lang' ( a selector of nikola translation ) and pases to a callable that expects a locale, we were using an implicit conversion.

We change that to an explicit conversion locale_from_lang(lang) -> locale , but in the linux codepath still we convert by identity, so it is a neutral change for linux.
In windows theres some translation, but before it was totally broken; still broken but better.

extracted logic in utils.Functionary.current_lang dealing with locales to guess_lang_from_current_locale