Add a note to CONTRIBUTING.rst
The main rationale is to have some niceties.  Commit messages are
the best place to describe big updates, and also small updates.

Note that nothing is going to enforce it (I’d probably have to build
some magic to check whether people do the imperative mood in the first
line anyways)

I also removed some trailing whitespace here or there (considering it
lights up in my vim¹ with a red at character, and I have a
semi-automated way of cleaning this up, I had to do this.)

¹ I didn’t do this earlier because I generally make simple edits
  using GitHub.com.

Signed-off-by: Chris “Kwpolska” Warrick <kwpolska@gmail.com>