missed an instance of the former setup in reST
Note that I am doing that ONLY because I want a single implementation
troughout the entire codebase, without using compiler-specific nonsense.

If we did not do this, we’d have to have “I can do it by myself” value
for the compiler.demote_headers variable (there is no AlreadyDone
singleton in Python, is there?  Lying with False does not count), and
we’d have to query the site config in the reST plugin.  Not worth the
hassle.

Signed-off-by: Chris “Kwpolska” Warrick <kwpolska@gmail.com>