Fix drafts being leaked in feeds.
This commit is a simple fix to work around the way `doit` seems to
handle file dependencies - if a file was a dependency for a task prior
to this run, but is no longer a dependency, the task is not run,
again. Only, the file dependencies in this run are checked for
changes.

We simply check use_in_feeds status for all the posts, to discover any
changes and regenerate the feed.

Fixes #934