Add `minify_lines` filter
For use with HTML, CSS, JS, XML, SVG, you name it.

Strips leading whitespace on every line as well as empty and all-
whitespace lines. Minimalist minifier that is very useful with the
output Nikola produces.

Has no whitespace intelligence, so that must be exercised by user.
Will break in situations where whitespace is required or expected,
like Python in <pre> or CDATA. Can be worked around using escaped
whitespace (&#32; for space and &#9; for tab.) This filter could be
expanded to do that as its own pre-processing  if there is user-
demand. With this in place, it could be turned on by default.