Drop Python 2.6 support.
Effective immediately, the Nikola project drops support for Python
version 2.6.  As a result, the current supported versions of Python
are 2.7, 3.3 and 3.4.  Users attempting to install with Python 2.7
will get an error message.  Users still on Python 2.6 should upgrade
to the newest version possible of the Python 2.7 branch, which is out
for more than 4 years (version 2.7.0 was released in July 2010)

Nikola v7.2.0 will continue to work with Python 2.6 — at least in
theory, as there probably are some obscure bugs and features that
don’t work.  We won’t provide support w/r/t 2.6-specific problems.

Signed-off-by: Chris Warrick <kwpolska@gmail.com>