fix #1522 -- hide untranslated posts’ translations
This hides the "Also available in:" text (and the rest of the translations
block) if there are no translations for this post.

Note that this uses `post.translated_to`.  Since `post.is_translation_available`
is just checking that list, we can check that list’s length and call it a day.
(if we were to display links to untranslated posts, it would break.)

cc @max-arnold

Signed-off-by: Chris Warrick <kwpolska@gmail.com>