add NikolaPygmentsHTML
This is a Nikola-specific Pygments HTML formatter, which produces code
identical to that of reST, but with `table` line numbers (if desired).
Code produced by the formatter is therefore compatible with existing
code.css files, and it also removes the need to write regexps to fix
things that didn’t comply.

The Markdown compiler still uses a regexp, another patch is needed for that.

Signed-off-by: Chris Warrick <kwpolska@gmail.com>