fix #1285 -- Add a modernized CodeBlock directive.
This is a saner and more modern rewrite of the CodeBlock directive
from reStructuredText.  This version defers the generation of all the
HTML to NikolaPygmentsHTML instead of a mix of a Pygments Lexer and a
custom hack.  As such, we can now use the `table` line numbers instead
of the broken `inline` mess that makes copy-pasting impossible.

Signed-off-by: Chris Warrick <kwpolska@gmail.com>