Automate releases!
This is an adaptation of my Python Project Template’s Release script
to Nikola’s needs. It automates two and a half sections from the task
list: housekeeping, versioning and tagging, and parts of website
update. It also uses source distributions instead of GitHub tarballs.
The updated checklist is 19 entries long instead of 39.

Signed-off-by: Chris Warrick <kwpolska@gmail.com>