$EDITOR error clarification
I believe the proposed error message is a bit more explicit.  When I first encountered this error as a user I was a bit confused.   I wasn't sure, at first, if $EDITOR was a Nikola setting.  Also, the second sentence wasn't clear on what "must be done manually ($EDITOR or the post)."  Semantics granted, but those were my personal first impressions and thought this might help make it clearer.  Also, this is my first ever attempt at a pull request.  :)

So I changed this:
"$EDITOR not set, cannot edit the post.  Please do it manually."

With this:
"$EDITOR not set within the operating system, cannot edit the post with \'-e\'.  Post must be altered manually."