Fail meaningfully for RSS_LINKS_APPEND_QUERY
Valid values for `FEED_LINKS_APPEND_QUERY` (or `RSS_LINKS_APPEND_QUERY`) are `None`, `False`, or a string. It's sort of devious, because presumably if False is allowed, True is also allowed. Further, the error is more buried since the config string is abstracted and passed down the stack, making it difficult to understand what is actually failing.

Perhaps there's a reasonable default that could be used in the truthy-true-but-not-string case, but having a descriptive failure is a step in that direction.

Tested locally with false/true/string.

Added myself to authors since this is about my fourth PR, I guess.