Remove empty return.
In `postlist.py` the `handler` function would return an empty string
and an empty value for `templatedata` if no posts were selected. This
caused the template to do no processing.

Removing this empty return lets the writer of a template properly use
the `if posts` dondition, and include an `else` if they need to.