Add enclosure assertions to gallery RSS test
The new test now makes assertions about the content of the enclosure (a
data structure about linked items, ie the images in this case) in each
RSS feed item.

This will enable us, in a later commit, to test the fallback to a
different MIME type when a file isn't recognized.