Use callback to set plugin module path for template location
This is a much cleaner (I think) approach to the problem of
figuring out the plugin's template path (see #3701). We use
yaspy's ability to provide a callback when loading the plugins.
The callback is passed a `yaspy.PluginInfo` instance, which
conveniently contains a reference to the plugin object itself,
and its filesystem path. So we can just add a method on our
plugin class to set the filesystem path as an attribute, and
have the callback function call that method.

Signed-off-by: Adam Williamson <awilliam@redhat.com>