Cleanup and fix loading of translations.
The new code clearly distinguishes between loading default
translations from the "base" theme and loading (additional)
translations from the (inherited) themes.

Also the former method put the directories onto sys.path and imported
the translation files as modules. This not only keeps the whole module
in memory (which might be neglectable nowadays) but also requires
extra code to reload the module for each theme.  The new method loads
the file and executes it using exec().

As a side-effect, this also fixed the bug that translations provided
by the "base" theme got lost if another theme provided additional
message strings.